Introduced Deploy Snapshot Target
As preparation for a more proper release management, introduced a new
target deploy-snapshot in build.xml. New target uses ivy:makepom and
ivy:publish tasks to upload a ready-built LAS2peer JAR to our archiva
installation in the snapshot folder. This is a preparation to have
automatic nightly builds with Jenkins.