CORS stuff for credentials when using UI on separate port
This is not a clean solution, but it works. With these settings, Chrome
will accept (and actually store) the session cookie and send the AJAX
requests without issues.
Should be used for development only for now. More research would be
needed on how to implement this properly to make it production-ready
(since CORS is a security concern ...).